+++ services/orchest-ctl/app/cli/main.py
@@ -7,7 +7,7 @@ import typer
 
 from app.orchest import OrchestApp
 from app.spec import get_container_config, inject_dict
-from app.cli import new_start as cli_start
+from app.cli import start as cli_start
 
 
 # TODO: utils.py
