--- services/orchest-ctl/app/connections.py
@@ -1,4 +1,0 @@
-from docker.client import DockerClient
-
-# Manage docker containers.
-docker_client = DockerClient.from_env()
